+++ app/src/main/java/com/axlecho/jtsviewer/network/JtsServerApi.java
@@ -41,7 +41,7 @@ public interface JtsServerApi {
     Observable<ResponseBody> getCollection();
 
     @GET("/collection/{id}/{page}")
-    Observable<ResponseBody> getCollectionDetail(@Path("id") int id, @Path("page") int page);
+    Observable<ResponseBody> getCollectionDetail(@Path("id") long id, @Path("page") int page);
 
     @FormUrlEncoded
     @POST("/member.php?mod=logging&action=login&loginsubmit=yes&handlekey=login&loginhash=LFR1d&inajax=1")
